requirements.txt
requests
beautifulsoup4
pyspark
pandas
sqlalchemy
psycopg2-binary
networkx
matplotlib